add support for literal function `collector` and `validator`
changes to lexer includes:
- feature in `INFORMATION` bit of `LITERAL` to hold index of associated
function.
- waypoint to implementation of collectors and validators

changes to exceptions include:
- macro definition of debug, ... for ease of use, (subjected to change
in future commits).

this commit changes interpretation of velidators as previously
validators were supposed to take in a const char* and ommit boolean if
given string is or is not a literal as expected, but in newer
interpretation it is supposed to be used as a tool for charachter
collection process, it now validates if given charachter is to be
considerd a part of literal or not.